+++ docs/RELEASE-NOTES.md
@@ -1,5 +1,17 @@
 # Release Notes
 
+## 1.1.0 - ?
+
+Breaking changes:
+- (None)
+
+New features:
+- Added `EnvironmentVariableProvider.GetLong`
+- Added `EnvironmentVariableProvider.GetLongOrDefault`
+
+Bug fixes / internal changes:
+- (None)
+
 ## 1.0.0 - 07 January 2021
 
 Initial version.